Dion Pushing
Adding a delay on the facial recognition. So the while the program is running is still detects motion but will only try to detect a face every ten second. Of course this can be altered to be longer or shorter time. This delay in the AI facial recognition also allows the camera to not lag as much, Since for every frame the there isn't a whole section of codes running. The program how does briefly freeze when a face is detected and will briefly show a red square if the face isn't recognized.